+++ apps/api/src/app.ts
@@ -14,6 +14,7 @@ import { activityRoutes } from "./activity";
 import { duosRoutes } from "./duos";
 import { announcementsRoutes } from "./announcements";
 import { wrappedRoutes } from "./wrappedRoutes";
+import { ideasRoutes } from "./ideas";
 
 export const app = new Hono();
 app.use("*", cors({ origin: WEB_ORIGINS, credentials: true }));
@@ -46,3 +47,4 @@ app.route("/", activityRoutes);
 app.route("/", duosRoutes);
 app.route("/", announcementsRoutes);
 app.route("/", wrappedRoutes);
+app.route("/", ideasRoutes);